Delete depositTicket operation + withdrawTicket models

+++ infrastructure/src/main/java/com/eu/gsys/wma/infrastructure/repositories/clients/CompanyClientRepository.java
@@ -3,5 +3,5 @@ package com.eu.gsys.wma.infrastructure.repositories.clients;
 import com.eu.gsys.wma.infrastructure.entities.clients.CompanyClientEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CompanyClientRepository extends CrudRepository<CompanyClientEntity, Integer> {
+public interface CompanyClientRepository extends CrudRepository<CompanyClientEntity, Long> {
 }
