+++ infrastructure/src/main/java/com/eu/gsys/wma/infrastructure/repositories/deposits/CompanyDepositRepository.java
@@ -4,7 +4,7 @@ import com.eu.gsys.wma.infrastructure.entities.clients.CompanyClientEntity;
 import com.eu.gsys.wma.infrastructure.entities.deposits.CompanyClientDepositEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CompanyDepositRepository extends CrudRepository<CompanyClientDepositEntity, Integer> {
+public interface CompanyDepositRepository extends CrudRepository<CompanyClientDepositEntity, Long> {
 
 	CompanyClientDepositEntity getDepositByClientEntity(CompanyClientEntity companyClientEntity);
 }
