+++ infrastructure/src/main/java/com/eu/gsys/wma/infrastructure/repositories/tickets/GristTicketRepository.java
@@ -3,5 +3,5 @@ package com.eu.gsys.wma.infrastructure.repositories.tickets;
 import com.eu.gsys.wma.infrastructure.entities.tickets.GristTicketEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GristTicketRepository extends CrudRepository<GristTicketEntity, Integer> {
+public interface GristTicketRepository extends CrudRepository<GristTicketEntity, Long> {
 }
